scExploreR config app

run_config() will launch a Shiny app used to configure datasets for use in the main browser.

Usage

run_config(./path_to_object.rds, ./path_to_config_file.yaml)

Arguments

object_path

relative or absolute path to the Seurat object to be configured. The file provided must be a .rds file created from a single Seurat object.

config_path

optional: if provided, the data from this file will be loaded when the user selects "load config file" in the config app. This should be a YAML file, though .rds files from versions 0.4.0 and earlier will also be accepted.

is_HDF5SummarizedExperiment

Set this to TRUE to load an HDF5-enabled SingleCellExperiment object saved via saveHDF5SummarizedExperiment. When loading an HDF5-enabled object, set the object_path to the directory of the HDF5-enabled object, created when saving the object via HDF5Array:saveHDF5SummarizedExperiment.

dev_mode

Used only for development. If TRUE, the server values for each option chosen by the user will be printed at the bottom of the "general" tab.